* In Room-Details-Request-Dto convert new-Price and old-Price to Double.

+++ src/main/java/com/AlTaraf/Booking/Payload/request/RoomDetails/RoomDetailsRequestDto.java
@@ -12,8 +12,8 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class RoomDetailsRequestDto {
     private int roomNumber;
-    private BigDecimal newPrice;
-    private BigDecimal oldPrice;
+    private Double newPrice;
+    private Double oldPrice;
     private int adultsAllowed;
     private int childrenAllowed;
 }
